Move data classes to correct package
