+++ src/jvmMain/kotlin/io/github/fourlastor/data/DemoData.kt
@@ -1,4 +1,4 @@
-package io.github.fourlastor.entity
+package io.github.fourlastor.data
 
 import java.io.File
