+++ src/jvmMain/kotlin/io/github/fourlastor/data/EntityUpdater.kt
@@ -1,3 +1,3 @@
-package io.github.fourlastor.entity
+package io.github.fourlastor.data
 
 typealias EntityUpdater = (id: Long, update: EntityUpdate) -> Unit
